+++ scripts/update.sh
@@ -18,6 +18,11 @@ ln -sf ~/.dotfiles/zsh/oh-my-zsh/materialshell-electro.zsh-theme ~/.oh-my-zsh/cu
 # Install homebrew, cask and app store apps
 brew bundle --file=~/.dotfiles/homebrew/Brewfile
 
+# Install Global npm packages
+for i in $(cat ~/.dotfiles/node/packages*); do
+    npm install -g $i
+done
+
 # Install Visual Studio Code Extensions
 for i in $(cat ~/.dotfiles/vscode/extensions*); do
     code --install-extension $i